Ignore commented code.
Added index map creation that stores the list of all the comments inside the textToSearch.  When the patternhits go against the textToSearch to look for replacements, they will consult the indexMap collection to see if the hit eg "long" was inside a comment block and if so, will leave it alone.

Given this code:

//as long as there are messages to process
long MessageId() { }

It will convert it this way:

//as long as there are messages to process
Int64 MessageId() { }

rather than what it does currently:

//as Int64 as there are messages to process
Int64 MessageId() { }